Analysis for the change in Salary
The data scientist indeed it does look like the maximum paid wage is going up over time.
Even though this line looks like it's steeper than the one above it, it actually is not significant. That's because again, the axis is not the same when there's two different graphs. It looks like for data scientist, the maximum paid wage is going up, but the median is not. For a software engineer, the maximum paid wage is going up over time, and the minimum paid wage is going down over time. Overall, again, the median paid wage is not changing. Data analyst can see that the minimum paid wage is going down overtime, whereas the maximum is staying basically the same. Again, the median is not changing. And for business analyst again, the minimum is going down,
7:02
the maximum is stranding up, but it's not significant. And the median paid wage is not changing at all. So according to this analysis it does look like the minimum salaries for Software Engineers, Data Analyst and Business Analyst are decreasing over time. Even though the median paid wages are remaining exactly the same for all of these different categories. The maximum paid wage for Data Scientists and Software Engineers on the other hand are increasing. So overall that suggests that there might be increasing amount of opportunity to be awarded with very high salaries in data scientist jobs and software engineering jobs, but overall the typical salaries for data related jobs seem to be staying quite constant. And for every category other than data scientists, as time goes on, there might actually be increased risk that you will make less than you expected.